Remove RubyTimeParsed#getNanoOfSecond, and change "nanoOfSecond" to int

+++ embulk-core/src/main/java/org/embulk/spi/time/RubyTimeParser.java
@@ -238,7 +238,7 @@ class RubyTimeParser
                             v = readDigitsMax();
                         }
 
-                        builder.setNanoOfSecond((!negative ? v : -v) * (int) Math.pow(10, 9 - (pos - initPos)));
+                        builder.setNanoOfSecond((int)(!negative ? v : -v) * (int) Math.pow(10, 9 - (pos - initPos)));
                         break;
                     }
                     case MINUTE_OF_HOUR: { // %M, %OM - Minute of the hour (00..59)
